Move classes to ForEachLoops namespace
