+++ src/ForEachLoops/Iterator.php
@@ -1,7 +1,7 @@
 <?php
 declare( strict_types = 1 );
 
-namespace PHP\Loops;
+namespace PHP\ForEachLoops;
 
 /**
  * Traverses an Iterable item in a foreach() loop, retrieving that item's key and value.
